Move chooser views from search into searchpromotions
