+++ .eslintrc.js
@@ -153,14 +153,14 @@ module.exports = {
       files: [
         'docs/_static/**',
         'wagtail/contrib/modeladmin/static_src/wagtailmodeladmin/js/prepopulate.js',
+        'wagtail/contrib/search_promotions/static_src/wagtailsearchpromotions/js/query-chooser-modal.js',
         'wagtail/contrib/search_promotions/templates/wagtailsearchpromotions/includes/searchpromotions_formset.js',
+        'wagtail/contrib/search_promotions/templates/wagtailsearchpromotions/queries/chooser_field.js',
         'wagtail/documents/static_src/wagtaildocs/js/add-multiple.js',
         'wagtail/embeds/static_src/wagtailembeds/js/embed-chooser-modal.js',
         'wagtail/images/static_src/wagtailimages/js/add-multiple.js',
         'wagtail/images/static_src/wagtailimages/js/focal-point-chooser.js',
         'wagtail/images/static_src/wagtailimages/js/image-url-generator.js',
-        'wagtail/search/static_src/wagtailsearch/js/query-chooser-modal.js',
-        'wagtail/search/templates/wagtailsearch/queries/chooser_field.js',
         'wagtail/snippets/static_src/wagtailsnippets/js/snippet-multiple-select.js',
         'wagtail/users/static_src/wagtailusers/js/group-form.js',
       ],
