+++ wagtail/contrib/search_promotions/.gitignore
