+++ wagtail/contrib/search_promotions/admin_urls.py
@@ -8,4 +8,10 @@ urlpatterns = [
     path("add/", views.add, name="add"),
     path("<int:query_id>/", views.edit, name="edit"),
     path("<int:query_id>/delete/", views.delete, name="delete"),
+    path("queries/chooser/", views.chooser, name="chooser"),
+    path(
+        "queries/chooser/results/",
+        views.chooserresults,
+        name="chooserresults",
+    ),
 ]
