+++ wagtail/contrib/search_promotions/static_src/wagtailsearchpromotions/js/query-chooser-modal.js
