+++ wagtail/contrib/search_promotions/templates/wagtailsearchpromotions/add.html
@@ -21,7 +21,7 @@
 
             <ul class="fields">
                 <li>
-                    {% include "wagtailsearch/queries/chooser_field.html" with field=query_form.query_string only %}
+                    {% include "wagtailsearchpromotions/queries/chooser_field.html" with field=query_form.query_string only %}
                 </li>
                 <li>
                     {% include "wagtailsearchpromotions/includes/searchpromotions_formset.html" with formset=searchpicks_formset only %}
@@ -39,12 +39,12 @@
 {% block extra_js %}
     {{ block.super }}
     {% include "wagtailadmin/pages/_editor_js.html" %}
-    <script src="{% versioned_static 'wagtailsearch/js/query-chooser-modal.js' %}"></script>
+    <script src="{% versioned_static 'wagtailsearchpromotions/js/query-chooser-modal.js' %}"></script>
     {{ form_media.js }}
 
     <script type="text/javascript">
         {% include "wagtailsearchpromotions/includes/searchpromotions_formset.js" with formset=searchpicks_formset only %}
-        {% include "wagtailsearch/queries/chooser_field.js" only %}
+        {% include "wagtailsearchpromotions/queries/chooser_field.js" only %}
 
         $(function() {
             createQueryChooser('{{ query_form.query_string.auto_id }}');
