+++ wagtail/contrib/search_promotions/templates/wagtailsearchpromotions/queries/chooser/results.html
