+++ wagtail/contrib/search_promotions/templates/wagtailsearchpromotions/queries/chooser_field.html
